Fix Design > Pages tests
